Create Mental Health - Classification Methods
The final model is the KNN model because the prediction accuracy is 0.89. 1NN has the highest prediction accuracy because it has the highest prediction accuracy out of all the 10 predictions. For example, 2NN to 10NN has decreasing prediction accuracy from 0.84 to 0.73, which means that 1NN has the highest prediction accuracy of 0.89. This means that we can accurately predict anxiety from the 5 variables correctly 89% of the time. 1NN means that the best value of K is 1. From the summary statistics in the logistic regression model, the null deviance is 128.21 and residual deviance is 119.77. Because these are high numbers, the model doesn’t explain the data well.

For the logistic regression model, the AIC is 140, which is a relatively large number. Therefore, because the value is so high, the model doesn't fit the data well. Additionally, all of the p-values are greater than 0.05, so the predictors aren't statistically significant. Additionally, the coefficients tell us that after accounting for all other variables, when Anxiety increases by a point, the odds of Age increase by a factor of 9.25, for example. Additionally, the coefficient of Age tells us that when increasing Age by a unit, the log odds of Anxiety are expected to decrease by  0.077, after holding other predictors fixed. 

For the classification tree, there are 4 internal nodes and 6 terminal nodes. GPA is the highest node, which means that GPA is the most important factor in determining anxiety than age and gender. To interpret the tree, if GPA is between 0-1.99, 2-2.50, and 3.50-4, the student is less than 18.5 years old, they’re female, and they’re less than 20.5 years old, then they don’t have anxiety. Additionally, if the student’s GPA is 2.50 - 3.00, they don’t have anxiety. 
Overall, the prediction accuracy is about 0.7. From the confusion matrix, the precision and recall are 0.68 and 1, respectively. When precision is 0.68, this means that when predicting Anxiety, the model is correct 68% of the time. Plus, when the recall is 1, it correctly identifies 100% of cases, or in other words, 100% of actual positives are identified correctly. This tells us that we can predict Anxiety from age, gender, GPA, and marital status reliably. 
In terms of implications for our target audience, which are young adults, we can confidently say that there’s a good chance that female students younger than 20.5 years old and has GPAs ranging from 0-1.99, 2.-2.5, and 3.5-4, do not have anxiety.